+++ cmd/gen-commands.ts
@@ -298,6 +298,7 @@ export const commands = [
     "CmdToggleTransparencyGrid", "Toggle Transparency Grid",
     "CmdTogglePageGrid", "Toggle Page Grid",
     "CmdConfigurePageGrid", "Configure Page Grid...",
+    "CmdTogglePdfAnnotationsToolbar", "Toggle PDF Annotations Toolbar",
     "CmdNone", "Do nothing",
 ];